Remove exploration steps
removing added exploratory steps as the increased entropy regularization
enables the agent to explore sufficently and learn to reach the target. The
exploratory steps do not appear to improve th eperformance beyond this.

Average episode rewards over 100 episodes running the deterministic policy for
the trained model without the exploratory steps is:

avg: 0.9972813845104688 max: 1.0000000000000002 min: 0.7281384510468834

and with the exploratory steps is:

avg: 0.991840809232942 max: 1.0000000000000002 min: 0.7280179918923811